Split init.lua to multiple files
Splitting init.lua for easier maintenance.
