+++ .config/nvim/lua/config/lazy.lua
@@ -1,0 +1,44 @@
+-- Bootstrap lazy.nvim {{{
+-- folke/lazy.nvim
+local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+if not (vim.uv or vim.loop).fs_stat(lazypath) then
+  local lazyrepo = "https://github.com/folke/lazy.nvim.git"
+  local out = vim.fn.system({ "git", "clone", "--filter=blob:none", "--branch=stable", lazyrepo, lazypath })
+  if vim.v.shell_error ~= 0 then
+    vim.api.nvim_echo({
+      { "Failed to clone lazy.nvim:\n", "ErrorMsg" },
+      { out, "WarningMsg" },
+      { "\nPress any key to exit..." },
+    }, true, {})
+    vim.fn.getchar()
+    os.exit(1)
+  end
+end
+vim.opt.rtp:prepend(lazypath)
+-- }}}
+
+-- Setup lazy.nvim
+-- lazy.nvim {{{
+require("lazy").setup({
+  spec = { import = "plugins" },
+  -- Other settings {{{
+  -- Configure any other settings here. See the documentation for more details.
+  -- colorscheme that will be used when installing plugins.
+  install = {
+    colorscheme = { "gruvbox-material" },
+    missing = true,
+  },
+  pkg = {
+    -- the first package source that is found for a plugin will be used.
+    sources = {
+      "lazy",
+      "rockspec", -- will only be used when rocks.enabled is true
+      "packspec",
+    },
+  },
+  rocks = {
+    enabled = false,
+  },
+  -- }}}
+})
+-- }}}
