+++ .config/nvim/lua/plugins/conform.lua
@@ -1,0 +1,64 @@
+return {
+  -- conform.nvim
+  -- code formatter
+  {
+    "stevearc/conform.nvim",
+    event = {},
+    cmd = { "ConformInfo" },
+    keys = {
+      {
+        -- Customize or remove this keymap to your liking
+        "<Leader>f",
+        function()
+          require("conform").format({ async = true })
+        end,
+        mode = "",
+        desc = "Format buffer",
+      },
+    },
+    -- This will provide type hinting with LuaLS
+    ---@module "conform"
+    ---@type conform.setupOpts
+    opts = {
+      -- Define your formatters
+      formatters_by_ft = {
+        lua = { "stylua" },
+        css = { "prettier" },
+        html = { "prettier" },
+        javascript = { "prettier" },
+        json = { "prettier" },
+        jsonc = { "prettier" },
+        markdown = { "prettier" },
+        python = { "ruff" },
+        sh = { "shfmt" },
+        bash = { "shfmt" },
+        tex = { "latexindent" },
+
+        -- Use the "*" filetype to run formatters on all filetypes.
+        -- ["*"] = { "codespell" },
+        -- Use the "_" filetype to run formatters on filetypes that don't
+        -- have other formatters configured.
+        ["_"] = { "trim_whitespace" },
+        -- You can now set a custom `lsp_format` option inside the "_" wildcard
+        -- itself. For example:
+        -- ["_"] = { "trim_whitespace", lsp_format = "last" },
+      },
+      -- Set default options
+      default_format_opts = {
+        lsp_format = "fallback",
+      },
+      -- Set up format-on-save
+      -- format_on_save = { timeout_ms = 500 },
+      -- Customize formatters
+      --   formatters = {
+      --     shfmt = {
+      --       prepend_args = { "-i", "2" },
+      --     },
+      --   },
+    },
+    -- init = function()
+    --   -- If you want the formatexpr, here is the place to set it
+    --   vim.o.formatexpr = "v:lua.require'conform'.formatexpr()"
+    -- end,
+  },
+}
