+++ .config/nvim/lua/plugins/gitsigns.lua
@@ -1,0 +1,10 @@
+return {
+  -- gitsigns.nvim
+  -- Git diff signs
+  {
+    "lewis6991/gitsigns.nvim",
+    config = function()
+      require("gitsigns").setup()
+    end,
+  },
+}
