+++ .config/nvim/lua/plugins/lualine.lua
@@ -1,0 +1,11 @@
+return {
+  -- lualine.nvim
+  -- Statusline/tabline
+  {
+    "nvim-lualine/lualine.nvim",
+    dependencies = { "nvim-tree/nvim-web-devicons" },
+    config = function()
+      require("lualine").setup()
+    end,
+  },
+}
