+++ .config/nvim/lua/plugins/mason-lspconfig.lua
@@ -1,0 +1,9 @@
+return {
+  -- mason-lspconfig.nvim
+  {
+    "mason-org/mason-lspconfig.nvim",
+    config = function()
+      require("mason-lspconfig").setup()
+    end,
+  },
+}
