+++ .config/nvim/lua/plugins/mason.lua
@@ -1,0 +1,13 @@
+return {
+  -- mason
+  {
+    "mason-org/mason.nvim",
+    config = function()
+      require("mason").setup({
+        ui = {
+          check_outdated_packages_on_open = false,
+        },
+      })
+    end,
+  },
+}
