+++ .config/nvim/lua/plugins/nvim-autopairs.lua
@@ -1,0 +1,10 @@
+return {
+  -- nvim-autopairs
+  {
+    "windwp/nvim-autopairs",
+    event = "InsertEnter",
+    config = true,
+    -- use opts = {} for passing setup options
+    -- this is equivalent to setup({}) function
+  },
+}
