+++ .config/nvim/lua/plugins/tagbar.lua
@@ -1,0 +1,11 @@
+return {
+  -- tagbar
+  -- Dispalys tags
+  {
+    "preservim/tagbar",
+    init = function()
+      vim.g.tagbar_width = 40
+      vim.keymap.set("n", "<Leader>tb", ":TagbarToggle<CR>", { noremap = true })
+    end,
+  },
+}
