+++ .config/nvim/lua/plugins/yazi.lua
@@ -1,0 +1,49 @@
+return {
+  -- yazi.nvim
+  -- yazi implementation to neovim
+  {
+    ---@type LazySpec
+    {
+      "mikavilpas/yazi.nvim",
+      event = "VeryLazy",
+      dependencies = {
+        { "nvim-lua/plenary.nvim", lazy = true },
+      },
+      keys = {
+        -- 👇 in this section, choose your own keymappings!
+        {
+          "<leader>-",
+          mode = { "n", "v" },
+          "<cmd>Yazi<cr>",
+          desc = "Open yazi at the current file",
+        },
+        {
+          -- Open in the current working directory
+          "<leader>cw",
+          "<cmd>Yazi cwd<cr>",
+          desc = "Open the file manager in nvim's working directory",
+        },
+        {
+          "<c-up>",
+          "<cmd>Yazi toggle<cr>",
+          desc = "Resume the last yazi session",
+        },
+      },
+      ---@type YaziConfig | {}
+      opts = {
+        -- if you want to open yazi instead of netrw, see below for more info
+        open_for_directories = false,
+        keymaps = {
+          show_help = "<f1>",
+        },
+      },
+      -- 👇 if you use `open_for_directories=true`, this is recommended
+      init = function()
+        -- mark netrw as loaded so it's not loaded at all.
+        --
+        -- More details: https://github.com/mikavilpas/yazi.nvim/issues/802
+        vim.g.loaded_netrwPlugin = 1
+      end,
+    },
+  },
+}
